Obsolete code removed

+++ config/constants.php
@@ -8,6 +8,6 @@ return [
     'ANIMATION_ONLY_TAG' => '#animationsOnly',
     'NO_TAGS_SEARCH_TAG' => '#noTags',
     'COMPRESSION_TAG' => '#compressedOnly',
-    'MAX_IMG_SIZE' => (int)(16 * 1024 * 1024), // 16 MB. Value is in bytes.
+    'MAX_IMG_SIZE' => 16 * 1024 * 1024, // 16 MB. Value is in bytes.
     'TMP_STORED_PREFIX' => '<fileStoredInTempDir>',
 ];
